feat(#629): moved component to common epic
